feat: wip for product variant, adding min order quantity column

+++ domain/Product/Actions/UpdateProductAction.php
@@ -39,29 +39,46 @@ class UpdateProductAction
         }
 
         if (count($productData->product_options)) {
-            foreach ($productData->product_options[0] as $productOption) {
-                $productOptionModel = ProductOption::findOrNew($productOption['id']);
-                $productOptionModel->name = $productOption['name'];
-                $productOptionModel->save();
+            foreach ($productData->product_options[0] as $key => &$productOption) {
+                $productOptionModel = ProductOption::find($productOption['id']);
 
-                foreach ($productOption['productOptionValues'] as $key => $productOptionValue) {
+                if ($productOptionModel) {
+                    $productOptionModel->product_id = $product->id;
+                    $productOptionModel->name = $productOption['name'];
+                    $productOptionModel->save();
+                } else {
+                    $newProductOptionModel = ProductOption::create([
+                        'product_id' => $product->id,
+                        'name' => $productOption['name'],
+                    ]);
+
+                    $productData->product_variants = $this->searchAndChangeValue(
+                        $productOption['id'],
+                        $productData->product_variants,
+                        $newProductOptionModel->id
+                    );
+                    $productData->product_options[0][$key]['id'] = $newProductOptionModel->id;
+                }
+
+                foreach ($productOption['productOptionValues'] as $key2 => $productOptionValue) {
                     $optionValueModel = ProductOptionValue::find($productOptionValue['id']);
 
                     if ($optionValueModel) {
                         $optionValueModel->name = $productOptionValue['name'];
-                        $optionValueModel->product_option_id = $productOptionValue['product_option_id'];
+                        $optionValueModel->product_option_id = $productOption['id'];
                         $optionValueModel->save();
                     } else {
                         $newOptionValueModel = ProductOptionValue::create([
                             'name' => $productOptionValue['name'],
-                            'product_option_id' => $productOptionValue['product_option_id'],
+                            'product_option_id' => $productOption['id'],
                         ]);
 
-                        $productOption['productOptionValues'][$key]['id'] = $newOptionValueModel->id;
+                        $productOption['productOptionValues'][$key2]['id'] = $newOptionValueModel->id;
                         $productData->product_variants = $this->searchAndChangeValue(
                             $productOptionValue['id'],
                             $productData->product_variants,
-                            $newOptionValueModel->id
+                            $newOptionValueModel->id,
+                            'option_value_id'
                         );
                     }
                 }
@@ -70,12 +87,29 @@ class UpdateProductAction
                 $mappedOptionValueIds = array_map(function ($item) {
                     return $item['id'];
                 }, $productOption['productOptionValues']);
-                $toRemoveOptionValues = ProductOptionValue::where(
-                    'product_option_id',
-                    $productOptionModel->id
-                )->whereNotIn('id', $mappedOptionValueIds)->get();
-                foreach ($toRemoveOptionValues as $optionValue) {
-                    $optionValue->delete();
+                if (count($mappedOptionValueIds)) {
+                    $toRemoveOptionValues = ProductOptionValue::where(
+                        'product_option_id',
+                        $productOption['id']
+                    )->whereNotIn('id', $mappedOptionValueIds)->get();
+                    foreach ($toRemoveOptionValues as $optionValue) {
+                        $optionValue->delete();
+                    }
+                }
+            }
+
+            // Removal of product options
+            $mappedOptionIds = array_map(function ($item) {
+                return $item['id'];
+            }, $productData->product_options[0]);
+
+            if (count($mappedOptionIds)) {
+                $toRemoveOptions = ProductOption::where(
+                    'product_id',
+                    $product->id
+                )->whereNotIn('id', $mappedOptionIds)->get();
+                foreach ($toRemoveOptions as $option) {
+                    $option->delete();
                 }
             }
         }
@@ -85,23 +119,37 @@ class UpdateProductAction
             $mappedVariantIds = array_map(function ($item) {
                 return $item['id'];
             }, $productData->product_variants);
-            $toRemoveProductVariants = ProductVariant::where('product_id', $product->id)
-                ->whereNotIn('id', $mappedVariantIds)->get();
-            foreach ($toRemoveProductVariants as $productVariant) {
-                $productVariant->delete();
+            if (count($mappedVariantIds)) {
+                $toRemoveProductVariants = ProductVariant::where('product_id', $product->id)
+                    ->whereNotIn('id', $mappedVariantIds)->get();
+
+                foreach ($toRemoveProductVariants as $productVariant) {
+                    $productVariant->delete();
+                }
             }
 
             foreach ($productData->product_variants as $productVariant) {
-                $productVariantModel = ProductVariant::findOrNew($productVariant['id']);
-
-                $productVariantModel->product_id = $product['id'];
-                $productVariantModel->sku = $productVariant['sku'];
-                $productVariantModel->combination = $productVariant['combination'];
-                $productVariantModel->retail_price = $productVariant['retail_price'];
-                $productVariantModel->selling_price = $productVariant['selling_price'];
-                $productVariantModel->stock = $productVariant['stock'];
-                $productVariantModel->status = $productVariant['status'];
-                $productVariantModel->save();
+                $productVariantModel = ProductVariant::find($productVariant['id']);
+                if ($productVariantModel) {
+                    $productVariantModel->product_id = $product['id'];
+                    $productVariantModel->sku = $productVariant['sku'];
+                    $productVariantModel->combination = $productVariant['combination'];
+                    $productVariantModel->retail_price = $productVariant['retail_price'];
+                    $productVariantModel->selling_price = $productVariant['selling_price'];
+                    $productVariantModel->stock = $productVariant['stock'];
+                    $productVariantModel->status = $productVariant['status'];
+                    $productVariantModel->save();
+                } else {
+                    ProductVariant::create([
+                        'product_id' => $product['id'],
+                        'sku' => $productVariant['sku'],
+                        'combination' => $productVariant['combination'],
+                        'retail_price' => $productVariant['retail_price'],
+                        'selling_price' => $productVariant['selling_price'],
+                        'stock' => $productVariant['stock'],
+                        'status' => $productVariant['status'],
+                    ]);
+                }
             }
         }
 
@@ -114,12 +162,12 @@ class UpdateProductAction
         return $product;
     }
 
-    private function searchAndChangeValue($needle, &$haystack, $newValue)
+    private function searchAndChangeValue($needle, &$haystack, $newValue, $field = 'option_id')
     {
         foreach ($haystack as $key => $variant) {
             foreach ($variant['combination'] as $key2 => $combination) {
-                if ($combination['option_value_id'] == $needle) {
-                    $haystack[$key]['combination'][$key2]['option_value_id'] = $newValue;
+                if ($combination[$field] == $needle) {
+                    $haystack[$key]['combination'][$key2][$field] = $newValue;
                 }
             }
         }
@@ -139,6 +187,7 @@ class UpdateProductAction
                 'weight' => $productData->weight,
                 'status' => $productData->status,
                 'stock' => $productData->stock,
+                'minimum_order_quantity' => $productData->minimum_order_quantity,
                 'is_digital_product' => $productData->is_digital_product,
                 'is_featured' => $productData->is_featured,
                 'is_special_offer' => $productData->is_special_offer,
